+++ CHANGES.md
@@ -188,6 +188,8 @@ To be released.
  -  Added `IStore.DeleteTxIdBlockHashIndex(TxId, BlockHash)` method.
     [[#1294], [#1328]]
  -  Added `IStore.IterateTxIdBlockHashIndex(TxId)` method.  [[#1294], [#1328]]
+ -  Parameter `miner` for `Block<T>()` and `Block<T>.Mine()` can no longer be
+    `null`.  [[#1341]]
  -  `Swarm<T>.StartAsync()` method became to receive `broadcastBlockInterval`
     (or `millisecondsBroadcastBlockInterval`) parameter.  [[#1351]]
  -  Added the parameter `minimumBroadcastTarget` to `NetMQTransport()`
@@ -359,6 +361,7 @@ To be released.
 [#1334]: https://github.com/planetarium/libplanet/pull/1334
 [#1339]: https://github.com/planetarium/libplanet/issues/1339
 [#1340]: https://github.com/planetarium/libplanet/pull/1340
+[#1341]: https://github.com/planetarium/libplanet/pull/1341
 [#1342]: https://github.com/planetarium/libplanet/pull/1342
 [#1343]: https://github.com/planetarium/libplanet/pull/1343
 [#1348]: https://github.com/planetarium/libplanet/pull/1348